refactor: move io in root of services
