+++ backend/app/services/utils/profiling/converter.py
@@ -1,7 +1,7 @@
 from app.services.rawdata.chk import CHK
 from app.services.rawdata.dat import DAT
 from app.services.rawdata.converter import MapConverter
-from app.services.mapdata.io import get_chkt
+from app.services.io import get_chkt
 from cProfile import Profile
 from io import BytesIO
 import pstats
